Replace crypto::symm with a rewritten version
